Crate nextest_runner

Core functionality for cargo nextest. For a higher-level overview, see that documentation.

For the basic flow of operations in nextest, see this blog post.

Modules

cargo_config

Support for emulating Cargo’s configuration file discovery.

config

Configuration support for nextest.

console

Experimental support for creating a Tokio console.

double_spawn

Support for double-spawning test processes.

errors

Errors produced by nextest.

indenter

Support for indenting multi-line displays.

list

Support for building and querying lists of test instances and test binaries.

partition

Support for partitioning test runs across several machines.

platform

Platform-related data structures.

reporter

Prints out and aggregates test execution statuses.

reuse_build

Reuse builds performed earlier.

runner

The test runner.

show_config

Functionality for showing configuration output of various kinds.

signal

Support for handling signals in nextest.

target_runner

Support for target runners

test_filter

Filtering tests based on user-specified parameters.

test_output

Utilities for capture output from tests run in a child process

update

Self-updates for nextest.

write_str

Support for string-only writes.