Module nextest_runner::cargo_config

Support for emulating Cargo’s configuration file discovery.

Since cargo config get is not stable as of Rust 1.61, nextest must do its own config file search.

Structs

CargoConfigs

A store for Cargo config files discovered from disk.

CargoEnvironmentVariable

An environment variable set in config.toml. See https://doc.rust-lang.org/cargo/reference/config.html#env

EnvironmentMap

Environment variables to set when running tests.

TargetTriple

Represents a target triple that’s being cross-compiled against.

Enums

CargoConfigSource

The source of a Cargo config.

TargetTripleSource

The place where a target triple’s configuration was picked up from.

Functions

relative_dir_for

Returns the directory against which relative paths are computed for the given config path.