Module nextest_runner::config

Configuration support for nextest.

Structs

ConfigIdentifier

An identifier used in configuration.

CustomTestGroup

Represents a custom test group.

FinalConfig

The state of nextest profiles after build platforms have been applied.

NextestConfig

Overall configuration for nextest.

NextestJunitConfig

JUnit configuration for nextest, returned by a NextestProfile.

NextestProfile

A configuration profile for nextest. Contains most configuration used by the nextest runner.

NextestVersionConfig

Nextest version configuration.

PreBuildPlatform

The state of nextest profiles before build platforms have been applied.

SlowTimeout

Type for the slow-timeout config key.

TestGroupConfig

Configuration for a test group.

TestSettings

Settings for individual tests.

ToolConfigFile

A tool-specific config file.

VersionOnlyConfig

A “version-only” form of the nextest configuration.

Enums

ConfigExperimental

Experimental configuration features.

NextestVersionEval

The result of checking whether a [NextestVersion] satisfies a requirement.

NextestVersionReq

Specification for a nextest version. Part of [NextestVersion].

RetryPolicy

Type for the retry config key.

TestGroup

Represents the test group a test is in.

TestThreads

Type for the test-threads config key.

ThreadsRequired

Type for the threads-required config key.

Functions

get_num_cpus

Gets the number of available CPUs and caches the value.