Module nextest_runner::errors

Errors produced by nextest.

Re-exports

pub use crate::reporter::structured::FormatVersionError;

Structs

ConfigParseCompiledDataError

An error that occurred while parsing config overrides or setup scripts.

ConfigParseError

An error that occurred while parsing the config.

InvalidConfigScriptName

The name of a configuration script is invalid (not a valid identifier).

InvalidCustomTestGroupName

The name of a test group is invalid (not a valid identifier).

PartitionerBuilderParseError

An error that occurs while parsing a PartitionerBuilder input.

ProfileNotFound

An error which indicates that a profile was requested but not known to nextest.

SignalHandlerSetupError

An error that occurred while setting up the signal handler.

TestThreadsParseError

Error returned while parsing a TestThreads value.

UnknownArchiveFormat

Represents an unknown archive format.

UnknownConfigScriptError

An unknown script was specified in the config.

UnknownHostPlatform

The host platform could not be determined.

UnknownTestGroupError

An unknown test group was specified in the config.

Enums

ArchiveCreateError

An error that occurs while archiving data.

ArchiveExtractError

An error occurred while extracting a file.

ArchiveReadError

An error occurred while reading a file.

CargoConfigError

An error occurred while constructing a CargoConfigs instance.

CollectTestOutputError

An error was returned while collecting test output.

ConfigParseErrorKind

The kind of error that occurred while parsing a config.

ConfigureHandleInheritanceError

An error occurred while configuring handles.

CreateTestListError

An error that occurs while parsing test list output.

FromMessagesError

An error that occurs in BinaryList::from_messages or RustTestArtifact::from_binary_list.

InvalidCargoCliConfigReason

The reason an invalid CLI config failed.

InvalidIdentifier

An identifier is invalid.

PathMapperConstructError

An error occurred in PathMapper::new.

PathMapperConstructKind

The kind of directory that failed to be read in PathMapper::new.

RustBuildMetaParseError

An error that occurs while parsing Rust build metadata from a summary.

ShowTestGroupsError

An error occurred while showing test groups.

TargetRunnerError

An error occurred determining the target runner

TargetTripleError

An error occurred while determining the cross-compiling target triple.

TestFilterBuilderError

An error that occurs while operating on a TestFilterBuilder.

TestRunnerBuildError

An error that occurs while building the test runner.

ToolConfigFileParseError

Error returned while parsing a ToolConfigFile value.

UpdateError

An error that occurs while performing a self-update.

UpdateVersionParseError

An error occurred while parsing an UpdateVersion.

WriteEventError

An error that occurs while writing an event.

WriteTestListError

An error that occurs while writing list output.

Functions

provided_by_tool

Returns the string provided by tool <tool>, if tool is Some.