Enum nextest_runner::errors::CargoConfigError

#[non_exhaustive]
pub enum CargoConfigError {
    GetCurrentDir(Error),
    CurrentDirInvalidUtf8(FromPathBufError),
    CliConfigParseError {
        config_str: String,
        error: TomlError,
    },
    CliConfigDeError {
        config_str: String,
        error: Error,
    },
    InvalidCliConfig {
        config_str: String,
        reason: InvalidCargoCliConfigReason,
    },
    NonUtf8Path(FromPathBufError),
    GetCargoHome(Error),
    FailedPathCanonicalization {
        path: Utf8PathBuf,
        error: Error,
    },
    ConfigReadError {
        path: Utf8PathBuf,
        error: Error,
    },
    ConfigParseError {
        path: Utf8PathBuf,
        error: Error,
    },
}

An error occurred while constructing a CargoConfigs instance.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

GetCurrentDir(Error)

Failed to retrieve the current directory.

CurrentDirInvalidUtf8(FromPathBufError)

The current directory was invalid UTF-8.

CliConfigParseError

Fields

config_str: String

The CLI config option.

error: TomlError

The error that occurred trying to parse the config.

Parsing a CLI config option failed.

CliConfigDeError

Fields

config_str: String

The CLI config option.

error: Error

The error that occurred trying to deserialize the config.

Deserializing a CLI config option into domain types failed.

InvalidCliConfig

Fields

config_str: String

The CLI config option.

reason: InvalidCargoCliConfigReason

The reason why this Cargo CLI config is invalid.

A CLI config option is not in the dotted key format.

NonUtf8Path(FromPathBufError)

A non-UTF-8 path was encountered.

GetCargoHome(Error)

Failed to retrieve the Cargo home directory.

FailedPathCanonicalization

Fields

path: Utf8PathBuf

The path that failed to canonicalize

error: Error

The error the occurred during canonicalization

Failed to canonicalize a path

ConfigReadError

Fields

path: Utf8PathBuf

The path of the config file

error: Error

The error that occurred trying to read the config file

Failed to read config file

ConfigParseError

Fields

path: Utf8PathBuf

The path of the config file

error: Error

The error that occurred trying to deserialize the config file

Failed to deserialize config file