Enum nextest_runner::errors::ConfigParseErrorKind

#[non_exhaustive]
pub enum ConfigParseErrorKind {

    BuildError(Box<ConfigError>),
    DeserializeError(Box<Error<ConfigError>>),
    VersionOnlyReadError(Error),
    VersionOnlyDeserializeError(Box<Error<Error>>),
    CompiledDataParseError(Vec<ConfigParseCompiledDataError>),
    InvalidTestGroupsDefined(BTreeSet<CustomTestGroup>),
    InvalidTestGroupsDefinedByTool(BTreeSet<CustomTestGroup>),
    UnknownTestGroups {
        errors: Vec<UnknownTestGroupError>,
        known_groups: BTreeSet<TestGroup>,
    },
    InvalidConfigScriptsDefined(BTreeSet<ScriptId>),
    InvalidConfigScriptsDefinedByTool(BTreeSet<ScriptId>),
    UnknownConfigScripts {
        errors: Vec<UnknownConfigScriptError>,
        known_scripts: BTreeSet<ScriptId>,
    },
    UnknownExperimentalFeatures {
        unknown: BTreeSet<String>,
        known: BTreeSet<ConfigExperimental>,
    },
    ExperimentalFeaturesInToolConfig {
        features: BTreeSet<String>,
    },
    ExperimentalFeatureNotEnabled {
        feature: ConfigExperimental,
    },

}

The kind of error that occurred while parsing a config.

Returned by ConfigParseError::kind.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BuildError(Box<ConfigError>)

An error occurred while building the config.

DeserializeError(Box<Error<ConfigError>>)

An error occurred while deserializing the config.

VersionOnlyReadError(Error)

An error occurred while reading the config file (version only).

VersionOnlyDeserializeError(Box<Error<Error>>)

An error occurred while deserializing the config (version only).

CompiledDataParseError(Vec<ConfigParseCompiledDataError>)

Errors occurred while parsing compiled data.

InvalidTestGroupsDefined(BTreeSet<CustomTestGroup>)

An invalid set of test groups was defined by the user.

InvalidTestGroupsDefinedByTool(BTreeSet<CustomTestGroup>)

An invalid set of test groups was defined by a tool config file.

UnknownTestGroups

Fields

errors: Vec<UnknownTestGroupError>

The list of errors that occurred.

known_groups: BTreeSet<TestGroup>

Known groups up to this point.

Some test groups were unknown.

InvalidConfigScriptsDefined(BTreeSet<ScriptId>)

An invalid set of config scripts was defined by the user.

InvalidConfigScriptsDefinedByTool(BTreeSet<ScriptId>)

An invalid set of config scripts was defined by a tool config file.

UnknownConfigScripts

Fields

errors: Vec<UnknownConfigScriptError>

The list of errors that occurred.

known_scripts: BTreeSet<ScriptId>

Known scripts up to this point.

Some config scripts were unknown.

UnknownExperimentalFeatures

Fields

unknown: BTreeSet<String>

The set of unknown features.

known: BTreeSet<ConfigExperimental>

The set of known features.

An unknown experimental feature or features were defined.

ExperimentalFeaturesInToolConfig

Fields

features: BTreeSet<String>

The name of the experimental feature.

A tool specified an experimental feature.

Tools are not allowed to specify experimental features.

ExperimentalFeatureNotEnabled

Fields

feature: ConfigExperimental

The feature that was not enabled.

An experimental feature was used but not enabled.