Module nextest_runner::reporter

Prints out and aggregates test execution statuses.

The main structure in this module is TestReporter.

Re-exports

pub use aggregator::heuristic_extract_description;

Modules

structured

Functionality for emitting structured, machine readable output in different formats

Structs

TestEvent

A test event.

TestReporter

Functionality to report test results to stderr, JUnit, and/or structured, machine-readable results to stdout

TestReporterBuilder

Test reporter builder.

Enums

CancelReason

The reason why a test run is being cancelled.

FinalStatusLevel

Status level to show at the end of test runs in the reporter output.

ReporterStderr

Standard error destination for the reporter.

StatusLevel

Status level to show in the reporter output.

TestEventKind

The kind of test event this is.

TestOutputDisplay

When to display test output in the reporter.