Enum nextest_runner::reporter::TestEvent

pub enum TestEvent<'a> {
    RunStarted {
        test_list: &'a TestList<'a>,
        run_id: Uuid,
    },
    TestStarted {
        test_instance: TestInstance<'a>,
        current_stats: RunStats,
        running: usize,
        cancel_state: Option<CancelReason>,
    },
    TestSlow {
        test_instance: TestInstance<'a>,
        retry_data: RetryData,
        elapsed: Duration,
        will_terminate: bool,
    },
    TestAttemptFailedWillRetry {
        test_instance: TestInstance<'a>,
        run_status: ExecuteStatus,
        delay_before_next_attempt: Duration,
        failure_output: TestOutputDisplay,
    },
    TestRetryStarted {
        test_instance: TestInstance<'a>,
        retry_data: RetryData,
    },
    TestFinished {
        test_instance: TestInstance<'a>,
        success_output: TestOutputDisplay,
        failure_output: TestOutputDisplay,
        junit_store_success_output: bool,
        junit_store_failure_output: bool,
        run_statuses: ExecutionStatuses,
        current_stats: RunStats,
        running: usize,
        cancel_state: Option<CancelReason>,
    },
    TestSkipped {
        test_instance: TestInstance<'a>,
        reason: MismatchReason,
    },
    RunBeginCancel {
        running: usize,
        reason: CancelReason,
    },
    RunPaused {
        running: usize,
    },
    RunContinued {
        running: usize,
    },
    RunFinished {
        run_id: Uuid,
        start_time: SystemTime,
        elapsed: Duration,
        run_stats: RunStats,
    },
}

A test event.

Events are produced by a TestRunner and consumed by a TestReporter.

Variants

RunStarted

Fields

test_list: &'a TestList<'a>

The list of tests that will be run.

The methods on the test list indicate the number of tests that will be run.

run_id: Uuid

The UUID for this run.

The test run started.

TestStarted

Fields

test_instance: TestInstance<'a>

The test instance that was started.

current_stats: RunStats

Current run statistics so far.

running: usize

The number of tests currently running, including this one.

cancel_state: Option<CancelReason>

The cancel status of the run. This is None if the run is still ongoing.

A test started running.

TestSlow

Fields

test_instance: TestInstance<'a>

The test instance that was slow.

retry_data: RetryData

Retry data.

elapsed: Duration

The amount of time that has elapsed since the beginning of the test.

will_terminate: bool

True if the test has hit its timeout and is about to be terminated.

A test was slower than a configured soft timeout.

TestAttemptFailedWillRetry

Fields

test_instance: TestInstance<'a>

The test instance that is being retried.

run_status: ExecuteStatus

The status of this attempt to run the test. Will never be success.

delay_before_next_attempt: Duration

The delay before the next attempt to run the test.

failure_output: TestOutputDisplay

Whether failure outputs are printed out.

A test attempt failed and will be retried in the future.

This event does not occur on the final run of a failing test.

TestRetryStarted

Fields

test_instance: TestInstance<'a>

The test instance that is being retried.

retry_data: RetryData

Data related to retries.

A retry has started.

TestFinished

Fields

test_instance: TestInstance<'a>

The test instance that finished running.

success_output: TestOutputDisplay

Test setting for success output.

failure_output: TestOutputDisplay

Test setting for failure output.

junit_store_success_output: bool

Whether the JUnit report should store success output for this test.

junit_store_failure_output: bool

Whether the JUnit report should store failure output for this test.

run_statuses: ExecutionStatuses

Information about all the runs for this test.

current_stats: RunStats

Current statistics for number of tests so far.

running: usize

The number of tests that are currently running, excluding this one.

cancel_state: Option<CancelReason>

The cancel status of the run. This is None if the run is still ongoing.

A test finished running.

TestSkipped

Fields

test_instance: TestInstance<'a>

The test instance that was skipped.

reason: MismatchReason

The reason this test was skipped.

A test was skipped.

RunBeginCancel

Fields

running: usize

The number of tests still running.

reason: CancelReason

The reason this run was canceled.

A cancellation notice was received.

RunPaused

Fields

running: usize

The number of tests currently running.

A SIGTSTP event was received and the run was paused.

RunContinued

Fields

running: usize

The number of tests that will be started up again.

A SIGCONT event was received and the run is being continued.

RunFinished

Fields

run_id: Uuid

The unique ID for this run.

start_time: SystemTime

The time at which the run was started.

elapsed: Duration

The amount of time it took for the tests to run.

run_stats: RunStats

Statistics for the run.

The test run finished.