Module nextest_runner::runner

The test runner.

The main structure in this module is TestRunner.

Structs

ExecuteStatus

Information about a single execution of a test.

ExecutionStatuses

Information about executions of a test, including retries.

RetryData

Data related to retries.

RunStats

Statistics for a test run.

TestRunner

Context for running tests.

TestRunnerBuilder

Test runner options.

Enums

AbortStatus

A regular exit code or Windows NT abort status for a test.

ExecutionDescription

A description of test executions obtained from ExecuteStatuses.

ExecutionResult

Whether a test passed, failed or an error occurred while executing the test.

Functions

configure_handle_inheritance

Configures stdout, stdin and stderr inheritance by test processes on Windows.