Enum nextest_runner::runner::ExecutionResult

pub enum ExecutionResult {
    Pass,
    Leak,
    Fail {
        abort_status: Option<AbortStatus>,
        leaked: bool,
    },
    ExecFail,
    Timeout,
}

Whether a test passed, failed or an error occurred while executing the test.

Variants

Pass

The test passed.

Leak

The test passed but leaked handles. This usually indicates that a subprocess that inherit standard IO was created, but it didn’t shut down when the test failed.

This is treated as a pass.

Fail

Fields

abort_status: Option<AbortStatus>

The abort status of the test, if any (for example, the signal on Unix).

leaked: bool

Whether a test leaked handles. If set to true, this usually indicates that a subprocess that inherit standard IO was created, but it didn’t shut down when the test failed.

The test failed.

ExecFail

An error occurred while executing the test.

Timeout

The test was terminated due to timeout.

Implementations

impl ExecutionResult

pub fn is_success(self) -> bool

Returns true if the test was successful.