Struct nextest_runner::target_runner::TargetRunner

pub struct TargetRunner { /* private fields */ }

A target runner used to execute a test binary rather than the default of executing natively.

Implementations

impl TargetRunner

pub fn new( configs: &CargoConfigs, build_platforms: &BuildPlatforms ) -> Result<Self, TargetRunnerError>

Acquires the target runner which can be set in a .cargo/config.toml or via a CARGO_TARGET_{TRIPLE}_RUNNER environment variable

pub fn empty() -> Self

Creates an empty target runner that does not delegate to any runner binaries.

pub fn target(&self) -> Option<&PlatformRunner>

Returns the target PlatformRunner.

pub fn host(&self) -> Option<&PlatformRunner>

Returns the host PlatformRunner.

pub fn for_build_platform( &self, build_platform: BuildPlatform ) -> Option<&PlatformRunner>

Returns the PlatformRunner for the given build platform (host or target).

pub fn all_build_platforms( &self ) -> [(BuildPlatform, Option<&PlatformRunner>); 2]

Returns the platform runners for all build platforms.