Trait nextest_runner::write_str::WriteStr

pub trait WriteStr {
    // Required methods
    fn write_str(&mut self, s: &str) -> Result<()>;
    fn write_str_flush(&mut self) -> Result<()>;

    // Provided methods
    fn write_char(&mut self, c: char) -> Result<()> { ... }
    fn write_fmt(&mut self, fmt: Arguments<'_>) -> Result<()> { ... }
}

A trait that abstracts over writing strings to a writer.

For more, see the module-level documentation.

Required Methods

fn write_str(&mut self, s: &str) -> Result<()>

Writes a string to the writer.

fn write_str_flush(&mut self) -> Result<()>

Flushes the writer, ensuring that all intermediately buffered contents reach their destination.

Provided Methods

fn write_char(&mut self, c: char) -> Result<()>

Writes a single character to the writer.

fn write_fmt(&mut self, fmt: Arguments<'_>) -> Result<()>

Writes a formatted string to the writer.