Module nextest_runner::errors

Errors produced by nextest.

Structs

ConfigParseError

An error that occurred while parsing the config.

JunitError

An error that occurred while producing JUnit XML.

OutputFormatParseError

An error that occurs while parsing an OutputFormat value from a string.

PartitionerBuilderParseError

An error that occurs while parsing a PartitionerBuilder input.

ProfileNotFound

An error which indicates that a profile was requested but not known to nextest.

RunIgnoredParseError

An error that occurs while parsing a RunIgnored value from a string.

StatusLevelParseError

Error returned while parsing a StatusLevel value from a string.

TestOutputDisplayParseError

Error returned while parsing a TestOutputDisplay value from a string.

Enums

FromMessagesError

An error that occurs in RustTestArtifact::from_messages.

ParseTestListError

An error that occurs while parsing test list output.

WriteEventError

An error that occurs while writing an event.

WriteTestListError

An error that occurs while writing list output.