Module nextest_runner::list

Support for building and querying lists of test instances and test binaries.

The main data structures in this module are:

TestList for test instances

BinaryList for test binaries

Structs

BinaryList

The list of Rust test binaries built by Cargo.

PathMapper

A helper for path remapping.

RustBuildMeta

Rust-related metadata used for builds and test runs.

RustTestArtifact

A Rust test binary built by Cargo. This artifact hasn’t been run yet so there’s no information about the tests within it.

RustTestBinary

A Rust test binary built by Cargo.

RustTestSuite

A suite of tests within a single Rust test binary.

TestInstance

Represents a single test with its associated binary.

TestList

List of test instances, obtained by querying the RustTestArtifact instances generated by Cargo.

Enums

BinaryListState

Typestate for BinaryList.

OutputFormat

Output formats for nextest.

SerializableFormat

A serialized, machine-readable output format.

TestListState

Typestate for TestList.