Enum nextest_runner::reporter::TestEvent

pub enum TestEvent<'a> {
    RunStarted {
        test_list: &'a TestList,
    },
    TestStarted {
        test_instance: TestInstance<'a>,
    },
    TestRetry {
        test_instance: TestInstance<'a>,
        run_status: TestRunStatus,
    },
    TestFinished {
        test_instance: TestInstance<'a>,
        run_statuses: RunStatuses,
    },
    TestSkipped {
        test_instance: TestInstance<'a>,
        reason: MismatchReason,
    },
    RunBeginCancel {
        running: usize,
        reason: CancelReason,
    },
    RunFinished {
        start_time: SystemTime,
        elapsed: Duration,
        run_stats: RunStats,
    },
}

Variants

RunStarted

The test run started.

Fields of RunStarted

test_list: &'a TestList

The list of tests that will be run.

The methods on the test list indicate the number of

TestStarted

A test started running.

Fields of TestStarted

test_instance: TestInstance<'a>

The test instance that was started.

TestRetry

A test failed and is being retried.

This event does not occur on the final run of a failing test.

Fields of TestRetry

test_instance: TestInstance<'a>

The test instance that is being retried.

run_status: TestRunStatus

The status of this attempt to run the test. Will never be success.

TestFinished

A test finished running.

Fields of TestFinished

test_instance: TestInstance<'a>

The test instance that finished running.

run_statuses: RunStatuses

Information about all the runs for this test.

TestSkipped

A test was skipped.

Fields of TestSkipped

test_instance: TestInstance<'a>

The test instance that was skipped.

reason: MismatchReason

The reason this test was skipped.

RunBeginCancel

A cancellation notice was received.

Fields of RunBeginCancel

running: usize

The number of tests still running.

reason: CancelReason

The reason this run was canceled.

RunFinished

The test run finished.

Fields of RunFinished

start_time: SystemTime

The time at which the run was started.

elapsed: Duration

The amount of time it took for the tests to run.

run_stats: RunStats

Statistics for the run.