Module nextest_runner::reuse_build

Reuse builds performed earlier.

Nextest allows users to reuse builds done on one machine. This module contains support for that.

The main data structures here are ReuseBuildInfo and PathMapper.

Structs

ArchiveReporter

Reporter for archive operations.

MetadataWithRemap

Metadata as either deserialized contents or a path, along with a possible directory remap.

PathMapper

A helper for path remapping.

ReuseBuildInfo

Reuse build information.

Enums

ArchiveEvent

An archive event.

ArchiveFormat

Archive format.

ExtractDestination

Where to extract a nextest archive to.

MetadataOrPath

Represents either a path to metadata or actual deserialized metadata.

Constants

BINARIES_METADATA_FILE_NAME

The name of the file in which binaries metadata is stored.

CARGO_METADATA_FILE_NAME

The name of the file in which Cargo metadata is stored.

Functions

archive_to_file

Archives test binaries along with metadata to the given file.